refactor(chronology): move the session modal view into event/
`tingle check` was red at +13. The residue was the seam handing the enroll
presenter's output to the modal, and it could not shrink: every candidate home
— the SessionData property it started as, and the view context it moved to —
sits inside `chronology/`, which is exactly the package `old-subdomain-loc`
counts. Trimming it further would have been shaping code around a counter.

So take the migration the counter is asking for instead. SessionModalComponentView
is self-contained — all four of its helpers are its own methods, and only urls.py
names it — and a modal on the event page belongs in `gates/web/django/event/`
next to the presenter it now calls. Moving it takes 96 lines out of the counted
package and turns the check from +13 into **-85 paid off**, with the seam left
exactly as the last review argued it should be.

Also merges origin/main, which had moved twice. The PAPERCUTS and .po conflicts
were both append-only; main's new papercut warns that a local `mise run messages`
strips `python-brace-format` flags, so those were checked explicitly and all
eleven survive.
